requirements.txt
streamlit
openai
whisper
torch
pydub
python-dotenv
langchain
transformer